add description

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,10 +1,15 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "lists.h"
+/**
+ * print_list - prints all elements of a list_t list
+ * @h: list to print
+ * Return: the number of nodes
+ */
 
 size_t print_list(const list_t *h)
 {
-	int counter = 0;
+	long unsigned int counter = 0;
 
 	if (h->str == NULL)
 	{
